Move fuzzing into ktool.kfuzz
